[refactor] file domain
