+++ src/main/java/flab/project/domain/file/enums/FileType.java
@@ -1,4 +1,4 @@
-package flab.project.data.enums;
+package flab.project.domain.file.enums;
 
 public enum FileType {
     PROFILE_IMAGE,
